Added doc to realizeType
